requirements.txt
jsonpickle==1.1
pillow==7.1.2
fastapi==0.54.2
uvicorn==0.11.5
requests==2.23.0
python-dotenv==0.13.0
beautifulsoup4==4.9.1
python-multipart==0.0.5
